[CRTool] move crtool Session handling out of the React app
